refactor(epic-plan): extract edit-flow into its own module for MI ceiling
Story #1499: the inlined runEditFlow/detectExistingSpec in epic-plan.js
pushed the wrapper's maintainability index below the close gate's
ceiling (-10pt vs epic baseline). Splitting the edit-in-place flow into
.agents/scripts/epic-plan-edit-flow.js keeps the wrapper a thin router
and lifts the MI back over the gate.
