+++ .agents/scripts/epic-plan.js
@@ -18,15 +18,11 @@
  * sub-CLI. This wrapper primarily owns the in-chat confirmation gate.
  */
 
-import { existsSync as defaultExistsSync } from 'node:fs';
 import { readFile } from 'node:fs/promises';
 import { parseArgs } from 'node:util';
 import { runDecomposePhase } from './epic-plan-decompose.js';
+import { detectExistingSpec, runEditFlow } from './epic-plan-edit-flow.js';
 import { runSpecPhase } from './epic-plan-spec.js';
-import {
-  confirmInteractive as defaultConfirmInteractive,
-  runReconcile,
-} from './epic-reconcile.js';
 import { runAsCli } from './lib/cli-utils.js';
 import {
   resolveConfig,
@@ -43,225 +39,12 @@ import {
   PLAN_PHASE_NAMES,
 } from './lib/orchestration/plan-runner/plan-router.js';
 import { createProvider } from './lib/provider-factory.js';
-import { specPath } from './lib/spec/loader.js';
 
-/**
- * Story #1499 / Task #1527 — Phase 2.5 edit-in-place entry-point detection.
- *
- * Resolve the on-disk spec file for `epicId` and report whether it
- * already exists. When present, the host LLM (this wrapper) routes the
- * plan invocation through the edit-in-place flow (`runEditFlow`) instead
- * of the author-then-reconcile (`runSpec` → `runDecompose`) chain.
- *
- * The check is intentionally a stat — neither YAML parse nor schema
- * validation runs here, so a malformed spec does not poison the
- * detection path. The downstream edit flow (Task #1530) calls `loadSpec`
- * to validate before doing any further work.
- *
- * Exported so tests can pin the routing predicate without re-deriving
- * the path convention from prose.
- *
- * @param {number|string} epicId
- * @param {{ existsSync?: typeof defaultExistsSync, specPathFn?: typeof specPath, epicsDir?: string }} [opts]
- * @returns {{ exists: boolean, path: string }}
- */
-export function detectExistingSpec(epicId, opts = {}) {
-  const existsFn = opts.existsSync ?? defaultExistsSync;
-  const specPathFn = opts.specPathFn ?? specPath;
-  const filePath = specPathFn(
-    epicId,
-    opts.epicsDir ? { epicsDir: opts.epicsDir } : {},
-  );
-  return { exists: existsFn(filePath), path: filePath };
-}
-
-/**
- * Edit-in-place flow for Phase 2.5 (Story #1499 / Task #1530).
- *
- * Reached when `detectExistingSpec` reports the spec is already on disk
- * for `epicId`. Wraps `epic-reconcile.js`'s `runReconcile` to:
- *
- *   1. Compute the structural plan (dry-run) and render it to stdout.
- *   2. Short-circuit with a no-changes message when the plan is empty —
- *      no operator prompt, no apply.
- *   3. Otherwise, gate the apply phase behind explicit operator
- *      confirmation (`y`/`yes`). When the operator declines, exit
- *      cleanly with `applied: false` and `reason: 'declined'`.
- *   4. On confirmation, invoke `runReconcile` a second time with
- *      `apply: true, yes: true` so the operator-confirmed apply runs
- *      end-to-end without prompting the embedded reconciler's own gate.
- *
- * The function is dependency-injection-heavy so tests can drive every
- * branch (dry-run-only, empty-diff, confirmed-apply, declined-apply)
- * without spawning a child process or hitting a real TTY.
- *
- * @param {{
- *   epicId: number,
- *   provider: object,
- *   specFilePath: string,
- *   apply?: boolean,
- *   reconcileFn?: typeof runReconcile,
- *   confirm?: typeof defaultConfirmInteractive,
- *   isTty?: () => boolean,
- *   stdout?: (line: string) => void,
- *   stderr?: (line: string) => void,
- *   loaderOpts?: object,
- * }} args
- * @returns {Promise<{
- *   epicId: number,
- *   mode: 'edit',
- *   specPath: string,
- *   applied: boolean,
- *   plan: object|null,
- *   exitCode: number,
- *   reason?: string,
- *   applyResult?: object,
- * }>}
- */
-export async function runEditFlow({
-  epicId,
-  provider,
-  specFilePath,
-  apply = false,
-  reconcileFn = runReconcile,
-  confirm = defaultConfirmInteractive,
-  isTty = () => Boolean(process.stdin.isTTY),
-  stdout = (line) => process.stdout.write(`${line}\n`),
-  stderr = (line) => process.stderr.write(`${line}\n`),
-  loaderOpts,
-}) {
-  Logger.info(
-    `[epic-plan] Existing spec detected for Epic #${epicId} at ${specFilePath}. Routing through edit-in-place flow.`,
-  );
-
-  // Step 1: dry-run to compute + render the plan. We always do this,
-  // regardless of `apply`, so the operator sees the diff before being
-  // prompted to confirm anything.
-  const dryRunResult = await reconcileFn(
-    {
-      epicId,
-      dryRun: true,
-      apply: false,
-      explicitDelete: false,
-      yes: false,
-    },
-    {
-      provider,
-      stdout,
-      stderr,
-      loaderOpts,
-    },
-  );
-
-  if (dryRunResult.exitCode !== 0) {
-    return {
-      epicId,
-      mode: 'edit',
-      specPath: specFilePath,
-      applied: false,
-      plan: dryRunResult.plan ?? null,
-      exitCode: dryRunResult.exitCode,
-      reason: 'dry-run-failed',
-    };
-  }
-
-  // Step 2: empty-plan short-circuit. The reconciler reports this via
-  // its `plan` envelope; we re-check rather than re-deriving the
-  // predicate so the two surfaces stay aligned.
-  const { isEmptyPlan } = await import(
-    './lib/orchestration/epic-spec-reconciler-ops.js'
-  );
-  if (dryRunResult.plan && isEmptyPlan(dryRunResult.plan)) {
-    stdout(
-      `[epic-plan] No structural changes detected for Epic #${epicId}. Spec is in sync with live state.`,
-    );
-    return {
-      epicId,
-      mode: 'edit',
-      specPath: specFilePath,
-      applied: false,
-      plan: dryRunResult.plan,
-      exitCode: 0,
-      reason: 'empty-diff',
-    };
-  }
-
-  // Step 3: apply path. The default is dry-run-only — the operator must
-  // pass `--apply` (CLI) or `apply: true` (programmatic) to opt in.
-  if (!apply) {
-    return {
-      epicId,
-      mode: 'edit',
-      specPath: specFilePath,
-      applied: false,
-      plan: dryRunResult.plan,
-      exitCode: 0,
-      reason: 'dry-run-only',
-    };
-  }
-
-  if (!isTty()) {
-    stderr(
-      '[epic-plan] --apply requires an interactive TTY for the operator confirmation gate.',
-    );
-    return {
-      epicId,
-      mode: 'edit',
-      specPath: specFilePath,
-      applied: false,
-      plan: dryRunResult.plan,
-      exitCode: 1,
-      reason: 'no-tty',
-    };
-  }
-
-  stdout(
-    `[epic-plan] Reviewed plan for Epic #${epicId}. Apply these structural changes?`,
-  );
-  const confirmed = await confirm();
-  if (!confirmed) {
-    stdout('[epic-plan] Edit-in-place declined by operator.');
-    return {
-      epicId,
-      mode: 'edit',
-      specPath: specFilePath,
-      applied: false,
-      plan: dryRunResult.plan,
-      exitCode: 0,
-      reason: 'declined',
-    };
-  }
-
-  // Step 4: confirmed apply. `yes: true` so we do not double-prompt
-  // through the embedded reconciler's gate — the operator already
-  // consented one layer up.
-  const applyResult = await reconcileFn(
-    {
-      epicId,
-      dryRun: false,
-      apply: true,
-      explicitDelete: false,
-      yes: true,
-    },
-    {
-      provider,
-      stdout,
-      stderr,
-      loaderOpts,
-    },
-  );
-
-  return {
-    epicId,
-    mode: 'edit',
-    specPath: specFilePath,
-    applied: applyResult.exitCode === 0,
-    plan: applyResult.plan ?? dryRunResult.plan,
-    exitCode: applyResult.exitCode,
-    applyResult: applyResult.applyResult,
-    reason: applyResult.exitCode === 0 ? 'applied' : 'apply-failed',
-  };
-}
+// Story #1499 / Task #1527: detect + edit-in-place flow live in their own
+// module (epic-plan-edit-flow.js) so this wrapper stays a thin router.
+// Re-exported for test ergonomics — `tests/scripts/epic-plan.edit-flow.test.js`
+// imports them through this surface.
+export { detectExistingSpec, runEditFlow };
 
 /**
  * Orchestrate the full local plan. Intentionally side-effect-free on its
